# openapi.yaml
openapi: 3.1.0
info:
  title: Local Printer API
  version: 0.1.0
  description: >
    REST API for managing uploads, previews, print submission, printer status,
    printer options, and CUPS jobs through a local CUPS queue.
servers:
  - url: http://192.168.100.99:8000
    description: Backend IP on LAN
  - url: http://ubuntu26-remote.local:8000
    description: Example LAN server hostname
  - url: http://drukarka.local:8000
    description: Alternative LAN server hostname
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create a local user account and session
      description: |
        Creates a LAN-local account, starts a session, and sets the
        `local_printer_session` HttpOnly cookie used by protected endpoints.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Account created and session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Start a session for an existing user
      description: |
        Validates credentials, starts a session, and sets the
        `local_printer_session` HttpOnly cookie used by protected endpoints.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
...
  /options:
    get:
      tags: [Printer]
      summary: Get detected printer capabilities
      operationId: getOptions
      parameters:
        - name: debug
          in: query
          required: false
          description: Include raw detected CUPS/PPD options when true.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Frontend-friendly detected printer options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrinterOptionsResponse"
        "503":
          description: CUPS unavailable or queue missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files:
    post:
      tags: [Files]
      summary: Upload a file for previewing or printing
      operationId: uploadFile
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File accepted and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "400":
          description: Invalid or corrupt file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/{file_id}/preview:
    get:
      tags: [Files]
      summary: Get preview metadata for an uploaded file
      operationId: getFilePreview
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Preview metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "404":
          description: Unknown file ID or file owned by another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/{file_id}/preview/{page}:
    get:
      tags: [Files]
      summary: Get one preview page as PNG
      operationId: getPreviewPage
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: page
          in: path
          required: true
          description: 1-based preview page number.
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Preview image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "400":
          description: Invalid page number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown file ID, file owned by another user, or preview page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /print:
    post:
      tags: [Print]
      summary: Submit an uploaded file to CUPS
      operationId: submitPrintJob
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintRequest"
      responses:
        "200":
          description: Print job submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "400":
          description: Invalid print request or invalid page range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown file ID or file owned by another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Printer or queue state prevents printing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: CUPS unavailable or printer queue unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs:
    get:
      tags: [Jobs]
      summary: List current user's CUPS jobs for the configured queue
      description: |
        Lists jobs from CUPS that match the current user's persisted or
        transient print submissions. The default `active` scope returns only
        not-completed jobs so frontend queue/task views do not show historical
        completed jobs as cancelable work. Jobs submitted by other users are
        filtered out before job metadata is returned.
      operationId: listJobs
      security:
        - SessionCookieAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          description: Which CUPS job set to return. Defaults to active.
          schema:
            type: string
            enum: [active, completed, all]
            default: active
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "400":
          description: Invalid job scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: CUPS unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /history:
    get:
      tags: [History]
      summary: List print history for the current user
      description: >-
        Refreshes nonterminal history from matching CUPS jobs and recovers saved
        purge markers. If CUPS is unavailable or a job disappears without a marker,
        returns its last known status.
      operationId: listHistory
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: Print history list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: local_printer_session
      description: HttpOnly session cookie set by signup/login.
...
  schemas:
...
    SignupRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: >
            Desired username. Usernames are normalized to lowercase and must be
            3-40 characters using letters, numbers, dot, underscore, or hyphen.
          example: alice
        password:
          type: string
          format: password
          minLength: 8
        display_name:
          type:
            - string
            - "null"
          maxLength: 80
          example: Alice
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Username is normalized to lowercase before lookup.
          example: alice
        password:
          type: string
          format: password
...
    AuthResponse:
      type: object
      required: [user, session]
      properties:
        user:
          $ref: "#/components/schemas/User"
        session:
          $ref: "#/components/schemas/AuthSession"
...
    MeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true